Use configurable logging to generate success and failure output only when an update is actually attempted.
Log ERROR messages on misconfiguration or failure when an update
attempt is made, an INFO message on success, and DEBUG messages for
everything else.

If we can't get the current IP due to connection error, assume
connection is unavailable and do nothing.

Allows users reduce verbosity when run from cron, to avoid sending
non-essential emails, or increase it when testing.

Remove redundant return statements from `check_ssl(). If the certificate
is not valid, `sys.exit()` gets called. If it is valid, we don't need to
return anything.

Remove default value (`False`) from `ip` argument to `update_ip_to_dns`
function. This function does not attempt to determine the current IP,
and the failure to do so is logged elsewhere.